requirements/dev.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astroid==3.0.2
asttokens==2.4.1
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.0
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.8
Django==5.0.1
django-allauth==0.60.0
django-cors-headers==4.3.1
django-coverage-plugin==3.1.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.3.1
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-stubs==3.14.5
djangorestframework_simplejwt==5.5.0
executing==2.0.1
factory-boy==3.3.0
Faker==22.2.0
filelock==3.13.1
flake8==7.0.0
flake8-isort==6.1.1
identify==2.5.33
idna==3.6
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.20.0
isort==5.13.2
jedi==0.19.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==25.0.0
pytest==7.4.4
pytest-django==4.7.0
pytest-sugar==0.9.7
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
ruff==0.11.7
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
termcolor==2.4.0
text-unidecode==1.3
tomlkit==0.12.3
traitlets==5.14.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240106
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.6.0
